backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
jinja2>=3.1.2
markdown>=3.5.0
requests>=2.31.0
beautifulsoup4>=4.12.0
feedparser>=6.0.10
crawl4ai>=0.3.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
aiofiles>=23.2.0
